fix sign-compare warning for helmholtz and tests

+++ src/helmholtz.cpp
@@ -211,12 +211,12 @@ namespace exafmm_t {
     real_t c[3] = {0.0};
     std::vector<RealVec> up_check_surf;
     up_check_surf.resize(MAXLEVEL+1);
-    for(size_t level = 0; level <= MAXLEVEL; level++) {
+    for(int level=0; level<=MAXLEVEL; level++) {
       up_check_surf[level].resize(NSURF*3);
       up_check_surf[level] = surface(P,c,2.95,level);
     }
     #pragma omp parallel for
-    for(int i=0; i<leafs.size(); i++) {
+    for(size_t i=0; i<leafs.size(); i++) {
       Node* leaf = leafs[i];
       int level = leaf->level;
       RealVec checkCoord(NSURF*3);
@@ -280,12 +280,12 @@ namespace exafmm_t {
     real_t c[3] = {0.0};
     std::vector<RealVec> dn_equiv_surf;
     dn_equiv_surf.resize(MAXLEVEL+1);
-    for(size_t level = 0; level <= MAXLEVEL; level++) {
+    for(int level = 0; level <= MAXLEVEL; level++) {
       dn_equiv_surf[level].resize(NSURF*3);
       dn_equiv_surf[level] = surface(P,c,2.95,level);
     }
     #pragma omp parallel for
-    for(int i=0; i<leafs.size(); i++) {
+    for(size_t i=0; i<leafs.size(); i++) {
       Node* leaf = leafs[i];
       int level = leaf->level;
       // down check surface potential -> equivalent surface charge
@@ -311,15 +311,15 @@ namespace exafmm_t {
     real_t c[3] = {0.0};
     std::vector<RealVec> dn_check_surf;
     dn_check_surf.resize(MAXLEVEL+1);
-    for(size_t level = 0; level <= MAXLEVEL; level++) {
+    for(int level = 0; level <= MAXLEVEL; level++) {
       dn_check_surf[level].resize(NSURF*3);
       dn_check_surf[level] = surface(P,c,1.05,level);
     }
     #pragma omp parallel for
-    for(int i=0; i<targets.size(); i++) {
+    for(size_t i=0; i<targets.size(); i++) {
       Node* target = &targets[i];
       NodePtrs& sources = target->P2L_list;
-      for(int j=0; j<sources.size(); j++) {
+      for(size_t j=0; j<sources.size(); j++) {
         Node* source = sources[j];
         RealVec trg_check_coord(NSURF*3);
         int level = target->level;
@@ -339,15 +339,15 @@ namespace exafmm_t {
     real_t c[3] = {0.0};
     std::vector<RealVec> up_equiv_surf;
     up_equiv_surf.resize(MAXLEVEL+1);
-    for(size_t level = 0; level <= MAXLEVEL; level++) {
+    for(int level=0; level<=MAXLEVEL; level++) {
       up_equiv_surf[level].resize(NSURF*3);
       up_equiv_surf[level] = surface(P,c,1.05,level);
     }
     #pragma omp parallel for
-    for(int i=0; i<targets.size(); i++) {
+    for(size_t i=0; i<targets.size(); i++) {
       Node* target = targets[i];
       NodePtrs& sources = target->M2P_list;
-      for(int j=0; j<sources.size(); j++) {
+      for(size_t j=0; j<sources.size(); j++) {
         Node* source = sources[j];
         RealVec sourceEquivCoord(NSURF*3);
         int level = source->level;
@@ -365,10 +365,10 @@ namespace exafmm_t {
   void P2P(NodePtrs& leafs) {
     NodePtrs& targets = leafs;   // assume sources == targets
     #pragma omp parallel for
-    for(int i=0; i<targets.size(); i++) {
+    for(size_t i=0; i<targets.size(); i++) {
       Node* target = targets[i];
       NodePtrs& sources = target->P2P_list;
-      for(int j=0; j<sources.size(); j++) {
+      for(size_t j=0; j<sources.size(); j++) {
         Node* source = sources[j];
         gradient_P2P(source->src_coord, source->src_value, target->trg_coord, target->trg_value);
       }
@@ -383,11 +383,11 @@ namespace exafmm_t {
     M2Ldata.resize(MAXLEVEL);
     // construct M2L target nodes for each level
     std::vector<NodePtrs> nodes_out(MAXLEVEL);
-    for(int i = 0; i < nonleafs.size(); i++) {
+    for(size_t i=0; i<nonleafs.size(); i++) {
       nodes_out[nonleafs[i]->level].push_back(nonleafs[i]);
     }
     // prepare for M2Ldata for each level
-    for(int l = 0; l < MAXLEVEL; l++) {
+    for(int l=0; l<MAXLEVEL; l++) {
       // construct M2L source nodes for current level
       std::set<Node*> nodes_in_;
       for(size_t i=0; i<nodes_out[l].size(); i++) {
@@ -472,7 +472,7 @@ namespace exafmm_t {
 
     for(size_t blk1=0; blk1<blk1_cnt; blk1++) {
     #pragma omp parallel for
-      for(size_t k=0; k<n3; k++) {
+      for(int k=0; k<n3; k++) {
         for(size_t mat_indx=0; mat_indx< mat_cnt; mat_indx++) {
           size_t interac_blk1 = blk1*mat_cnt+mat_indx;
           size_t interaction_count_offset0 = (interac_blk1==0?0:interaction_count_offset[interac_blk1-1]);
@@ -524,14 +524,14 @@ namespace exafmm_t {
       complex_t* up_equiv = &all_up_equiv[fft_offset[node_idx]];  // offset ptr of node's 8 child's up_equiv in all_up_equiv, size=8*NSURF
       real_t* up_equiv_f = &fft_in[fftsize*node_idx];   // offset ptr of node_idx in fft_in vector, size=fftsize
 
-      for(size_t k=0; k<NSURF; k++) {
+      for(int k=0; k<NSURF; k++) {
         size_t idx = map[k];
-        for(int j0=0; j0<(int)NCHILD; j0++)
+        for(int j0=0; j0<NCHILD; j0++)
           equiv_t[idx+j0*n3] = up_equiv[j0*NSURF+k];
       }
       fft_execute_dft(plan, reinterpret_cast<fft_complex*>(&equiv_t[0]), (fft_complex*)&buffer[0]);
-      for(size_t j=0; j<n3; j++) {
-        for(size_t k=0; k<NCHILD; k++) {
+      for(int j=0; j<n3; j++) {
+        for(int k=0; k<NCHILD; k++) {
           up_equiv_f[2*(NCHILD*j+k)+0] = buffer[2*(n3*k+j)+0];
           up_equiv_f[2*(NCHILD*j+k)+1] = buffer[2*(n3*k+j)+1];
         }
@@ -568,13 +568,13 @@ namespace exafmm_t {
       ComplexVec buffer1(NCHILD*n3, 0);
       real_t* dn_check_f = &fft_out[fftsize*node_idx];  // offset ptr for node_idx in fft_out vector, size=fftsize
       complex_t* dn_equiv = &all_dn_equiv[ifft_offset[node_idx]];  // offset ptr for node_idx's child's dn_equiv in all_dn_equiv, size=numChilds * NSURF
-      for(size_t j=0; j<n3; j++)
-        for(size_t k=0; k<NCHILD; k++) {
+      for(int j=0; j<n3; j++)
+        for(int k=0; k<NCHILD; k++) {
           buffer0[2*(n3*k+j)+0] = dn_check_f[2*(NCHILD*j+k)+0];
           buffer0[2*(n3*k+j)+1] = dn_check_f[2*(NCHILD*j+k)+1];
         }
       fft_execute_dft(plan, (fft_complex*)&buffer0[0], reinterpret_cast<fft_complex*>(&buffer1[0]));
-      for(size_t k=0; k<NSURF; k++) {
+      for(int k=0; k<NSURF; k++) {
         size_t idx = map[k];
         for(int j0=0; j0<NCHILD; j0++)
           dn_equiv[NSURF*j0+k]+=buffer1[idx+j0*n3];
@@ -590,14 +590,14 @@ namespace exafmm_t {
     ComplexVec all_up_equiv(numNodes*NSURF);
     ComplexVec all_dn_equiv(numNodes*NSURF);
     #pragma omp parallel for collapse(2)
-    for(int i=0; i<numNodes; i++) {
+    for(size_t i=0; i<numNodes; i++) {
       for(int j=0; j<NSURF; j++) {
         all_up_equiv[i*NSURF+j] = nodes[i].up_equiv[j];
         all_dn_equiv[i*NSURF+j] = nodes[i].dn_equiv[j];
       }
     }
     size_t fftsize = 2 * 8 * n3;
-    for(int l = 0; l < MAXLEVEL; l++) {
+    for(int l=0; l<MAXLEVEL; l++) {
       AlignedVec fft_in(M2Ldata[l].fft_offset.size()*fftsize, 0.);
       AlignedVec fft_out(M2Ldata[l].ifft_offset.size()*fftsize, 0.);
       fft_up_equiv(M2Ldata[l].fft_offset, all_up_equiv, fft_in);
@@ -605,7 +605,7 @@ namespace exafmm_t {
       ifft_dn_check(M2Ldata[l].ifft_offset, fft_out, all_dn_equiv);
     }
     #pragma omp parallel for collapse(2)
-    for(int i=0; i<numNodes; i++) {
+    for(size_t i=0; i<numNodes; i++) {
       for(int j=0; j<NSURF; j++) {
         nodes[i].up_equiv[j] = all_up_equiv[i*NSURF+j];
         nodes[i].dn_equiv[j] = all_dn_equiv[i*NSURF+j];
